pyproject.toml
[tool.poetry]
name = "bird-inference-api"
version = "0.1.0"
description = "API de inferencia de aves com FastAPI e TensorFlow"
authors = ["Dattaflow <dev@dattaflow.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.1"
uvicorn = { version = "0.29.0", extras = ["standard"] }
python-dotenv = "1.0.1"
pydantic = "2.7.1"
pandas = "2.2.2"
numpy = "1.26.4"
tensorflow = "2.16.1"
pillow = "10.3.0"
python-multipart = "0.0.9"
typing-extensions = ">=4.10.0"
boto3 = "1.34.131"
botocore = "1.34.131"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
